Support strict dependencies thanks to `accept`/`reject` primitives
This commit introduces support for strict dependencies. Strict dependencies are constraints on the
versions of dependencies in the graph. When a dependency version is strict, it is expected that no
other dependency disagrees with it. It means that:

- if `A` says `strictly 1.0`, but `B` says `1.1`, then the build will fail
- if `A` says `strictly [1.0, 2.0)`, and `B` says `1.1`, then `1.1` will be selected
- if `A` says `strictly 1.1`, but `B` says `1.0`, then `1.1` is selected

Implementation-wise, a dependency version is now represented as a `ModuleVersionConstraint`, which
will internally be exposed as 2 different primitives:
   - `preferred`, a `VersionSelector` that selects the _preferred version_ (or baseline) of a dependency
   - `reject`, a `VersionSelector` which allows _rejecting_ specific versions

This commit does **not** address publication of such constraints.